// app/src/app/shared/dto/model/createEventDto.ts
/**
 * EventManagement.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateEventDto { 
    title?: string | null;
    description?: string | null;
    date?: string;
    maxCapacity?: number;
}