// app/src/app/shared/dto/model/updateEventDto.ts
/**
 * EventManagement.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateEventDto { 
    title?: string | null;
    description?: string | null;
    date?: string;
    maxCapacity?: number;
}